+++ src/plugins/chat/bot_config.toml
@@ -39,7 +39,7 @@ read_allowed = [
     752426484,#nd1
     115843978,#nd2
     # 168718420 #bh
-    912378791
+    # 912378791
 ]
 
 talk_allowed = [
@@ -60,14 +60,14 @@ talk_allowed = [
     # 168718420#bh
     # 752426484,#nd1
     # 115843978,#nd2
-    912378791
+    # 912378791
 ]
 
 talk_frequency_down = [
     549292720, #mrfz
     435591861, #m43black
     # 231561425,
-    975992476,
+    # 975992476,
     1140700103,
     # 534940728
     # 752426484,#nd1
